requirements.txt
requests
python-dotenv
fastapi
uvicorn
sqlalchemy
psycopg2-binary
email-validator
python-jose[cryptography]
passlib[bcrypt]
python-multipart

